+++ layer-atlas/src/main/java/com/layer/atlas/adapters/AtlasMessagesAdapter.java
@@ -43,14 +43,14 @@ import java.util.Set;
  * it can render, can create new View hierarchies for its Message types, and can render (bind)
  * Message data with its created View hierarchies.  Typically, CellFactories are segregated by
  * MessagePart MIME types (e.g. "text/plain", "image/jpeg", and "application/vnd.geo+json").
- *
+ * <p>
  * Under the hood, the AtlasMessagesAdapter is a RecyclerView.Adapter, which automatically recycles
  * its list items within view-type "buckets".  Each registered CellFactory actually creates two such
  * view-types: one for cells sent by the authenticated user, and another for cells sent by remote
  * actors.  This allows the AtlasMessagesAdapter to efficiently render images sent by the current
  * user aligned on the left, and images sent by others aligned on the right, for example.  In case
  * this sent-by distinction is of value when rendering cells, it provided as the `isMe` argument.
- *
+ * <p>
  * When rendering Messages, the AtlasMessagesAdapter first determines which CellFactory to handle
  * the Message with calling CellFactory.isBindable() on each of its registered CellFactories. The
  * first CellFactory to return `true` is used for that Message.  Then, the adapter checks for
@@ -89,10 +89,13 @@ public class AtlasMessagesAdapter extends RecyclerView.Adapter<AtlasMessagesAdap
 
     private Integer mRecipientStatusPosition;
 
-    //Stye
+    //Style
     private MessageStyle mMessageStyle;
 
     private RecyclerView mRecyclerView;
+    private boolean mReadReceiptsEnabled = true;
+
+    protected boolean mShouldShowAvatarInOneOnOneConversations;
 
     public AtlasMessagesAdapter(Context context, LayerClient layerClient, Picasso picasso) {
         mLayerClient = layerClient;
@@ -175,6 +178,33 @@ public class AtlasMessagesAdapter extends RecyclerView.Adapter<AtlasMessagesAdap
 
     public View getFooterView() {
         return mFooterView;
+    }
+
+    /**
+     * @return If the Avatar for the other participant in a one on one conversation  will be shown
+     * or not
+     */
+    public boolean getShouldShowAvatarInOneOnOneConversations() {
+        return mShouldShowAvatarInOneOnOneConversations;
+    }
+
+    /**
+     * @param shouldShowAvatarInOneOnOneConversations Whether the Avatar for the other participant
+     *                                                in a one on one conversation should be shown
+     *                                                or not
+     */
+    public void setShouldShowAvatarInOneOnOneConversations(boolean shouldShowAvatarInOneOnOneConversations) {
+        this.mShouldShowAvatarInOneOnOneConversations = shouldShowAvatarInOneOnOneConversations;
+    }
+
+    /**
+     * Set whether or not the conversation supports read receipts. This determines if the read
+     * receipts should be shown in the view holders.
+     *
+     * @param readReceiptsEnabled true if the conversation is adapter is used for supports read receipts
+     */
+    public void setReadReceiptsEnabled(boolean readReceiptsEnabled) {
+        mReadReceiptsEnabled = readReceiptsEnabled;
     }
 
 
@@ -319,7 +349,9 @@ public class AtlasMessagesAdapter extends RecyclerView.Adapter<AtlasMessagesAdap
                 viewHolder.mCell.setAlpha(1.0f);
             }
         } else {
-            message.markAsRead();
+            if (mReadReceiptsEnabled) {
+                message.markAsRead();
+            }
             // Sender name, only for first message in cluster
             if (!oneOnOne && (cluster.mClusterWithPrevious == null || cluster.mClusterWithPrevious == ClusterType.NEW_SENDER)) {
                 Identity sender = message.getSender();
@@ -338,13 +370,17 @@ public class AtlasMessagesAdapter extends RecyclerView.Adapter<AtlasMessagesAdap
 
             // Avatars
             if (oneOnOne) {
-                // Not in one-on-one conversations
-                viewHolder.mAvatar.setVisibility(View.GONE);
+                if (mShouldShowAvatarInOneOnOneConversations) {
+                    viewHolder.mAvatar.setVisibility(View.VISIBLE);
+                    viewHolder.mAvatar.setParticipants(message.getSender());
+
+                } else {
+                    viewHolder.mAvatar.setVisibility(View.GONE);
+                }
             } else if (cluster.mClusterWithNext == null || cluster.mClusterWithNext != ClusterType.LESS_THAN_MINUTE) {
                 // Last message in cluster
                 viewHolder.mAvatar.setVisibility(View.VISIBLE);
                 viewHolder.mAvatar.setParticipants(message.getSender());
-
                 // Add the position to the positions map for Identity updates
                 mIdentityEventListener.addIdentityPosition(position, Collections.singleton(message.getSender()));
             } else {
@@ -376,7 +412,7 @@ public class AtlasMessagesAdapter extends RecyclerView.Adapter<AtlasMessagesAdap
     }
 
     private void updateViewHolderForRecipientStatus(CellViewHolder viewHolder, int position, Message message) {
-        if (mRecipientStatusPosition != null && mRecipientStatusPosition == position) {
+        if (mReadReceiptsEnabled && mRecipientStatusPosition != null && mRecipientStatusPosition == position) {
             int readCount = 0;
             boolean delivered = false;
             Map<Identity, Message.RecipientStatus> statuses = message.getRecipientStatus();
@@ -442,7 +478,6 @@ public class AtlasMessagesAdapter extends RecyclerView.Adapter<AtlasMessagesAdap
         if (!(viewHolder instanceof CellViewHolder)) return null;
         return ((CellViewHolder) viewHolder).mMessage;
     }
-
 
     //==============================================================================================
     // Clustering
@@ -521,11 +556,13 @@ public class AtlasMessagesAdapter extends RecyclerView.Adapter<AtlasMessagesAdap
     //==============================================================================================
 
     private void updateRecipientStatusPosition() {
-        Integer oldPosition = mRecipientStatusPosition;
-        // Set new position to last in the list
-        mRecipientStatusPosition = mQueryController.getItemCount() - 1;
-        if (oldPosition != null) {
-            notifyItemChanged(oldPosition);
+        if (mReadReceiptsEnabled) {
+            Integer oldPosition = mRecipientStatusPosition;
+            // Set new position to last in the list
+            mRecipientStatusPosition = mQueryController.getItemCount() - 1;
+            if (oldPosition != null) {
+                notifyItemChanged(oldPosition);
+            }
         }
     }